+++ pytsite/assetman/__init__.py
@@ -3,7 +3,7 @@
 # Public API
 from . import _error as error
 from ._api import register_package, add, remove, dump_js, dump_css, url, add_inline, dump_inline, get_urls, \
-    get_locations, reset, detect_collection, build, is_package_registered
+    get_locations, reset, detect_collection, build, is_package_registered, register_global
 
 __author__ = 'Alexander Shepetko'
 __email__ = 'a@shepetko.com'